FriendListAdatper.kt -> onLongClick 구현

+++ app/src/main/java/com/example/solaroid/datasource/album/AlbumDataSource.kt
@@ -25,7 +25,7 @@ class AlbumDataSource {
                             name = hashMap["name"] as String,
                             thumbnail = hashMap["thumbnail"] as String,
                             participants = hashMap["participants"] as String,
-                            numOfParticipants = hashMap["numOfParticipants"] as Int,
+                            numOfParticipants = hashMap["numOfParticipants"] as Long,
                             key = hashMap["key"] as String
                         ).asDatabaseModel(user)
