+++ app/src/main/java/com/example/solaroid/datasource/album/RequestAlbumDataSource.kt
@@ -20,7 +20,7 @@ class RequestAlbumDataSource {
                         name = hashMap["name"] as String,
                         thumbnail = hashMap["thumbnail"] as String,
                         participants = hashMap["participants"] as String,
-                        numOfParticipants = hashMap["numOfParticipants"] as Int,
+                        numOfParticipants =   hashMap["numOfParticipants"] as Long,
                         albumKey = hashMap["albumKey"] as String,
                         key = hashMap["key"] as String
                     ).asDomainModel()
