+++ app/src/main/java/com/example/solaroid/models/firebase/AlbumModel.kt
@@ -13,7 +13,7 @@ data class FirebaseAlbum(
     val name: String,
     val thumbnail: String,
     val participants:String,
-    val numOfParticipants: Int,
+    val numOfParticipants: Long,
     val key: String
 ) {
 }
@@ -26,7 +26,7 @@ fun FirebaseAlbum.asDatabaseModel(user:String) : DatabaseAlbum {
         name,
         thumbnail,
         participants,
-        numOfParticipants,
+        numOfParticipants.toInt(),
         key,
         user
     )
@@ -39,7 +39,7 @@ data class FirebaseRequestAlbum(
     val name: String,
     val thumbnail: String,
     val participants: String,
-    val numOfParticipants: Int,
+    val numOfParticipants: Long,
     val albumKey: String,
     val key: String
     ) {
@@ -52,7 +52,7 @@ fun FirebaseRequestAlbum.asDomainModel() : RequestAlbum {
         name,
         BitmapUtils.stringToBitmap(thumbnail),
         participants,
-        numOfParticipants,
+        numOfParticipants.toInt(),
         albumKey,
         key
     )
